pyproject.toml
[project]
name = "hybridppo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gymnasium[mujoco,other]>=1.2.3",
    "minari[hdf5,hf]>=0.5.3",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "scipy>=1.16.3",
    "stable-baselines3>=2.7.1",
    "tensorboard>=2.20.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "wandb>=0.23.1",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["hybridppo*"]
